improved Client

+++ SnakeClient/SnakeClient/Position.cs
@@ -20,7 +20,7 @@ namespace SnakeClient
             Y = y;
         }
 
-        public bool IsInBounds(int x1, int y1, int x2, int y2)
+        public bool IsInBounds(float x1, float y1, float x2, float y2)
         {
             if(X >= x1 && Y >= y1 && X <= x2 && Y <= y2)
                 return true;
